now passing type of invoice to fetch function
